backend/pyproject.toml
[project]
name = "hormonaiq-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "supabase>=2.9",
    "google-genai>=0.8",
    "weasyprint>=62",
    "jinja2>=3.1",
    "pydantic-settings>=2.4",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
